+++ src/GuiQt/QGLWidgetTextRenderer.cxx
@@ -793,12 +793,12 @@ QGLWidgetTextRenderer::getBoundsForTextInModelSpace(const AnnotationText& /*anno
                                                     const float /*modelSpaceScaling*/,
                                                     const float /*heightOrWidthForPercentageSizeText*/,
                                                     const DrawingFlags& /*flags*/,
-                                                    float* /*bottomLeftOut[3]*/,
-                                                    float* /*bottomRightOut[3]*/,
-                                                    float* /*topRightOut[3]*/,
-                                                    float* /*topLeftOut[3]*/,
-                                                    float* /*underlineStartOut[3]*/,
-                                                    float* /*underlineEndOut[3]*/)
+                                                    double* /*bottomLeftOut[3]*/,
+                                                    double* /*bottomRightOut[3]*/,
+                                                    double* /*topRightOut[3]*/,
+                                                    double* /*topLeftOut[3]*/,
+                                                    double* /*underlineStartOut[3]*/,
+                                                    double* /*underlineEndOut[3]*/)
 {
     
 }
