+++ src/GuiQt/QGLWidgetTextRenderer.h
@@ -78,12 +78,12 @@ namespace caret {
                                                   const float modelSpaceScaling,
                                                   const float heightOrWidthForPercentageSizeText,
                                                   const DrawingFlags& flags,
-                                                  float bottomLeftOut[3],
-                                                  float bottomRightOut[3],
-                                                  float topRightOut[3],
-                                                  float topLeftOut[3],
-                                                  float underlineStartOut[3],
-                                                  float underlineEndOut[3]) override;
+                                                  double bottomLeftOut[3],
+                                                  double bottomRightOut[3],
+                                                  double topRightOut[3],
+                                                  double topLeftOut[3],
+                                                  double underlineStartOut[3],
+                                                  double underlineEndOut[3]) override;
         
         virtual void getBoundsForTextAtViewportCoords(const AnnotationText& annotationText,
                                                       const BrainOpenGLTextRenderInterface::DrawingFlags& flags,
